seprate cypress func
